requirements.txt
alembic==1.4.2
Babel==2.8.0
blinker==1.4
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
dominate==2.5.1
Flask==1.1.2
Flask-Babel==1.0.0
Flask-Bootstrap4==4.0.2
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.3
Flask-WTF==0.14.3
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
mysqlclient==1.4.6
python-dateutil==2.8.1
python-dotenv==0.13.0
python-editor==1.0.4
python-slugify==4.0.0
pytz==2020.1
requests==2.23.0
six==1.15.0
SQLAlchemy==1.3.17
text-unidecode==1.3
urllib3==1.25.9
visitor==0.1.3
Werkzeug==1.0.1
WTForms==2.3.1
